+++ demisto_sdk/commands/common/content/objects/pack_objects/pack.py
@@ -25,8 +25,6 @@ from demisto_sdk.commands.common.content.objects.pack_objects import (
     Widget)
 from demisto_sdk.commands.common.content.objects_factory import \
     path_to_pack_object
-from demisto_sdk.commands.create_artifacts.content_artifacts_creator import \
-    ArtifactsManager
 from wcmatch.pathlib import Path
 
 logger: logging.Logger
@@ -209,11 +207,12 @@ class Pack:
 
         return obj
 
-    def sign_pack(self, dumped_pack_dir: Path, signature_string: str):
+    def sign_pack(self, dumped_pack_dir: Path, sign_directory: Path, signature_string: str):
         """ Signs pack folder and creates signature file.
 
         Args:
             dumped_pack_dir (Path): Path to the updated pack to sign.
+            sign_directory (Path): Path to the signDirectory executable file.
             signature_string (str): Base64 encoded string used to sign the pack.
 
         """
@@ -222,9 +221,9 @@ class Pack:
         try:
             with open('keyfile', 'wb') as keyfile:
                 keyfile.write(signature_string.encode())
-            command = f'./signDirectory {dumped_pack_dir} keyfile base64'
+            full_command = f'{sign_directory} {dumped_pack_dir} keyfile base64'
 
-            signing_process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
+            signing_process = subprocess.Popen(full_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
             output, err = signing_process.communicate()
             signing_process.wait()
 
@@ -238,18 +237,18 @@ class Pack:
         except Exception as error:
             logger.error(f'Error while trying to sign pack {self.path.name}.\n {error}')
 
-    def encrypt_pack(self, artifact_manager: ArtifactsManager, zip_pack_path: Path, encryption_key: str):
+    def encrypt_pack(self, zip_pack_path: Path, encryptor: Path, encryption_key: str):
         """ Encrypt pack zip.
 
         Args:
-            artifact_manager (ArtifactsManager): Artifacts manager object.
             zip_pack_path (Path): Path to the pack not encrypted script.
+            encryptor (Path): Path to the encryptor executable file.
             encryption_key (str): The encryption key for the packs.
 
         """
         try:
             output_file = str(zip_pack_path).replace('_not_encrypted.zip', '.zip')
-            full_command = f'{artifact_manager.encryptor} {zip_pack_path} {output_file} "{encryption_key}"'
+            full_command = f'{encryptor} {zip_pack_path} {output_file} "{encryption_key}"'
             encryption_process = subprocess.Popen(full_command, stdout=subprocess.PIPE, stderr=subprocess.PIPE,
                                                   shell=True)
             output, err = encryption_process.communicate()
