+++ demisto_sdk/commands/common/content/objects/pack_objects/pack_metadata/pack_metadata.py
@@ -4,14 +4,13 @@ import os
 from datetime import datetime
 from typing import Dict, List, Union
 
+from demisto_sdk.commands.common import tools
 from demisto_sdk.commands.common.constants import (PACKS_PACK_META_FILE_NAME,
                                                    XSOAR_AUTHOR, XSOAR_SUPPORT,
                                                    XSOAR_SUPPORT_URL,
                                                    ContentItems)
 from demisto_sdk.commands.common.content.objects.abstract_objects import \
     JSONObject
-from demisto_sdk.commands.create_artifacts.content_artifacts_creator import \
-    CORE_PACKS_LIST
 from demisto_sdk.commands.find_dependencies.find_dependencies import \
     PackDependencies
 from packaging.version import Version, parse
@@ -23,6 +22,7 @@ DATETIME_FORMAT = '%Y-%m-%dT%H:%M:%SZ'
 PARTNER_SUPPORT = 'partner'
 XSOAR_CERTIFIED = 'certified'
 XSOAR_EULA_URL = 'https://github.com/demisto/content/blob/master/LICENSE'
+CORE_PACKS_LIST = tools.get_remote_file('Tests/Marketplace/core_packs_list.json') or []
 
 
 class PackMetaData(JSONObject):
@@ -569,7 +569,6 @@ class PackMetaData(JSONObject):
 
         except Exception:
             logger.error(f'Failed loading {pack_name} user metadata.')
-            return None
 
     def handle_dependencies(self, pack_name: str, id_set_path: str) -> None:
         """Updates pack's dependencies using the find_dependencies command.
