+++ demisto_sdk/commands/common/content/tests/objects/pack_objects/pack_metadata/pack_metadata_test.py
@@ -1,4 +1,7 @@
+import os
+from contextlib import contextmanager
 from datetime import datetime
+from shutil import rmtree
 
 import pytest
 from demisto_sdk.commands.common.constants import (PACKS_DIR, XSOAR_AUTHOR,
@@ -8,11 +11,33 @@ from demisto_sdk.commands.common.content.objects.pack_objects import \
     PackMetaData
 from demisto_sdk.commands.common.content.objects_factory import \
     path_to_pack_object
+from demisto_sdk.commands.common.logger import logging_setup
 from demisto_sdk.commands.common.tools import src_root
+from packaging.version import parse
+from TestSuite.test_tools import ChangeCWD
 
 TEST_DATA = src_root() / 'tests' / 'test_files'
 TEST_CONTENT_REPO = TEST_DATA / 'content_slim'
 PACK_METADATA = TEST_CONTENT_REPO / PACKS_DIR / 'Sample01' / 'pack_metadata.json'
+UNIT_TEST_DATA = (src_root() / 'commands' / 'create_artifacts' / 'tests' / 'data')
+
+
+@contextmanager
+def temp_dir():
+    """Create Temp directory for test.
+
+     Open:
+        - Create temp directory.
+
+    Close:
+        - Delete temp directory.
+    """
+    temp = UNIT_TEST_DATA / 'temp'
+    try:
+        temp.mkdir(parents=True, exist_ok=True)
+        yield temp
+    finally:
+        rmtree(temp)
 
 
 def test_objects_factory():
@@ -136,3 +161,229 @@ def test_dump_with_price(mocker):
     mocker.patch.object(json, 'dump', side_effect=mock_json_dump)
 
     obj.dump('metadata_file')
+
+
+def test_load_user_metadata_basic(repo):
+    """
+    When:
+        - Dumping a specific pack, processing the pack's metadata.
+
+    Given:
+        - Pack object.
+
+    Then:
+        - Verify that pack's metadata information was loaded successfully.
+
+    """
+    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (ArtifactsManager)
+
+    pack_1 = repo.setup_one_pack('Pack1')
+    pack_1.pack_metadata.write_json(
+        {
+            'name': 'Pack Number 1',
+            'description': 'A description for the pack',
+            'created': '2020-06-08T15:37:54Z',
+            'price': 0,
+            'support': 'xsoar',
+            'url': 'some url',
+            'email': 'some email',
+            'currentVersion': '1.1.1',
+            'author': 'Cortex XSOAR',
+            'tags': ['tag1'],
+            'dependencies': [{'dependency': {'dependency': '1'}}]
+        }
+    )
+
+    with ChangeCWD(repo.path):
+        with temp_dir() as temp:
+            artifact_manager = ArtifactsManager(artifacts_path=temp,
+                                                content_version='6.0.0',
+                                                zip=False,
+                                                suffix='',
+                                                cpus=1,
+                                                packs=True)
+
+    artifact_manager.content.packs['Pack1'].metadata.load_user_metadata('Pack1', 'Pack Number 1', pack_1.path)
+    result = artifact_manager.content.packs['Pack1'].metadata
+    assert result.id == 'Pack1'
+    assert result.name == 'Pack Number 1'
+    assert result.description == 'A description for the pack'
+    assert result.created == datetime(2020, 6, 8, 15, 37, 54)
+    assert result.price == 0
+    assert result.support == 'xsoar'
+    assert result.url == 'some url'
+    assert result.email == 'some email'
+    assert result.certification == 'certified'
+    assert result.current_version == parse('1.1.1')
+    assert result.author == 'Cortex XSOAR'
+    assert result.tags == ['tag1']
+    assert result.dependencies == [{'dependency': {'dependency': '1'}}]
+
+
+def test_load_user_metadata_advanced(repo):
+    """
+    When:
+        - Dumping a specific pack, processing the pack's metadata.
+
+    Given:
+        - Pack object.
+
+    Then:
+        - Verify that pack's metadata information was loaded successfully.
+
+    """
+    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (ArtifactsManager)
+
+    pack_1 = repo.setup_one_pack('Pack1')
+    pack_1.pack_metadata.write_json(
+        {
+            'name': 'Pack Number 1',
+            'price': 10,
+            'tags': ['tag1'],
+            'useCases': ['usecase1'],
+            'vendorId': 'vendorId',
+            'vendorName': 'vendorName'
+        }
+    )
+
+    with ChangeCWD(repo.path):
+        with temp_dir() as temp:
+            artifact_manager = ArtifactsManager(artifacts_path=temp,
+                                                content_version='6.0.0',
+                                                zip=False,
+                                                suffix='',
+                                                cpus=1,
+                                                packs=True)
+
+    artifact_manager.content.packs['Pack1'].metadata.load_user_metadata('Pack1', 'Pack Number 1', pack_1.path)
+    result = artifact_manager.content.packs['Pack1'].metadata
+    assert result.id == 'Pack1'
+    assert result.name == 'Pack Number 1'
+    assert result.price == 10
+    assert result.vendor_id == 'vendorId'
+    assert result.vendor_name == 'vendorName'
+    assert result.tags == ['tag1', 'Use Case']
+
+
+def test_load_user_metadata_no_metadata_file(repo, capsys):
+    """
+    When:
+        - Dumping a pack with no pack_metadata file.
+
+    Given:
+        - Pack object.
+
+    Then:
+        - Verify that exceptions are written to the logger.
+
+    """
+    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (ArtifactsManager)
+    import demisto_sdk.commands.common.content.objects.pack_objects.pack_metadata as metadata_class
+
+    metadata_class.logger = logging_setup(3)
+
+    pack_1 = repo.setup_one_pack('Pack1')
+    pack_1.pack_metadata.write_json(
+        {
+            'name': 'Pack Number 1',
+            'price': 'price',
+            'tags': ['tag1'],
+            'useCases': ['usecase1'],
+            'vendorId': 'vendorId',
+            'vendorName': 'vendorName'
+        }
+    )
+
+    with ChangeCWD(repo.path):
+        os.remove(pack_1.pack_metadata.path)
+        with temp_dir() as temp:
+            artifact_manager = ArtifactsManager(artifacts_path=temp,
+                                                content_version='6.0.0',
+                                                zip=False,
+                                                suffix='',
+                                                cpus=1,
+                                                packs=True)
+
+    artifact_manager.content.packs['Pack1'].metadata.load_user_metadata('Pack1', 'Pack Number 1', pack_1.path)
+
+    captured = capsys.readouterr()
+    assert 'Pack1 pack is missing pack_metadata.json file.' in captured.err
+
+
+def test_load_user_metadata_invalid_price(repo, capsys):
+    """
+    When:
+        - Dumping a pack with invalid price in pack_metadata file.
+
+    Given:
+        - Pack object.
+
+    Then:
+        - Verify that exceptions are written to the logger.
+
+    """
+    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (ArtifactsManager)
+    import demisto_sdk.commands.common.content.objects.pack_objects.pack_metadata as metadata_class
+
+    metadata_class.logger = logging_setup(3)
+
+    pack_1 = repo.setup_one_pack('Pack1')
+    pack_1.pack_metadata.write_json(
+        {
+            'name': 'Pack Number 1',
+            'price': 'price',
+            'tags': ['tag1'],
+            'useCases': ['usecase1'],
+            'vendorId': 'vendorId',
+            'vendorName': 'vendorName'
+        }
+    )
+
+    with ChangeCWD(repo.path):
+        with temp_dir() as temp:
+            artifact_manager = ArtifactsManager(artifacts_path=temp,
+                                                content_version='6.0.0',
+                                                zip=False,
+                                                suffix='',
+                                                cpus=1,
+                                                packs=True)
+
+    artifact_manager.content.packs['Pack1'].metadata.load_user_metadata('Pack1', 'Pack Number 1', pack_1.path)
+
+    captured = capsys.readouterr()
+    assert 'Pack Number 1 pack price is not valid. The price was set to 0.' in captured.err
+
+
+def test_load_user_metadata_bad_pack_metadata_file(repo, capsys):
+    """
+    When:
+        - Dumping a pack with invalid pack_metadata file.
+
+    Given:
+        - Pack object.
+
+    Then:
+        - Verify that exceptions are written to the logger.
+
+    """
+    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (ArtifactsManager)
+    import demisto_sdk.commands.common.content.objects.pack_objects.pack_metadata as metadata_class
+
+    metadata_class.logger = logging_setup(3)
+
+    pack_1 = repo.setup_one_pack('Pack1')
+    pack_1.pack_metadata.write_as_text('Invalid of course {')
+
+    with ChangeCWD(repo.path):
+        with temp_dir() as temp:
+            artifact_manager = ArtifactsManager(artifacts_path=temp,
+                                                content_version='6.0.0',
+                                                zip=False,
+                                                suffix='',
+                                                cpus=1,
+                                                packs=True)
+
+    artifact_manager.content.packs['Pack1'].metadata.load_user_metadata('Pack1', 'Pack Number 1', pack_1.path)
+
+    captured = capsys.readouterr()
+    assert 'Failed loading Pack1 user metadata.' in captured.err
