+++ demisto_sdk/commands/common/content/tests/objects/pack_objects/pack_test.py
@@ -1,3 +1,6 @@
+from contextlib import contextmanager
+from pathlib import Path
+from shutil import rmtree
 from typing import Tuple
 
 import pytest
@@ -8,11 +11,31 @@ from demisto_sdk.commands.common.content.objects.pack_objects import (
     IncidentType, IndicatorField, IndicatorType, Integration, LayoutsContainer,
     PackIgnore, PackMetaData, Playbook, Readme, ReleaseNote, Report, Script,
     SecretIgnore, Widget)
+from demisto_sdk.commands.common.logger import logging_setup
 from demisto_sdk.commands.common.tools import src_root
 
 TEST_DATA = src_root() / 'tests' / 'test_files'
 TEST_CONTENT_REPO = TEST_DATA / 'content_slim'
 PACK = TEST_CONTENT_REPO / PACKS_DIR / 'Sample01'
+UNIT_TEST_DATA = (src_root() / 'commands' / 'create_artifacts' / 'tests' / 'data')
+
+
+@contextmanager
+def temp_dir():
+    """Create Temp directory for test.
+
+     Open:
+        - Create temp directory.
+
+    Close:
+        - Delete temp directory.
+    """
+    temp = UNIT_TEST_DATA / 'temp'
+    try:
+        temp.mkdir(parents=True, exist_ok=True)
+        yield temp
+    finally:
+        rmtree(temp)
 
 
 @pytest.mark.parametrize(argnames="attribute, content_type, items",
@@ -55,3 +78,194 @@ def test_generators_detection(attribute: str, content_type: Tuple[type], items: 
 def test_detection(attribute: str, content_type: type):
     pack = Pack(PACK)
     assert isinstance(pack.__getattribute__(attribute), content_type)
+
+
+def test_sign_pack_exception_thrown(repo, capsys, mocker):
+    """
+    When:
+        - Signing a pack.
+
+    Given:
+        - Pack object.
+        - No signing executable.
+
+    Then:
+        - Verify that exceptions are written to the logger.
+
+    """
+    import demisto_sdk.commands.common.content.objects.pack_objects.pack as pack_class
+    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
+    import subprocess
+
+    mocker.patch.object(subprocess, 'Popen', autospec=True)
+
+    pack_class.logger = logging_setup(3)
+
+    pack = repo.create_pack('Pack1')
+    content_object_pack = Pack(pack.path)
+    signer_path = Path('./signer')
+
+    content_object_pack.sign_pack(content_object_pack.path, signer_path, 'key')
+
+    captured = capsys.readouterr()
+    assert 'Error while trying to sign pack Pack1' in captured.err
+
+
+def test_sign_pack_error_from_subprocess(repo, capsys, fake_process):
+    """
+    When:
+        - Signing a pack.
+
+    Given:
+        - Pack object.
+        - subprocess is failing due to an error.
+
+    Then:
+        - Verify that exceptions are written to the logger.
+
+    """
+    import demisto_sdk.commands.common.content.objects.pack_objects.pack as pack_class
+    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
+
+    pack_class.logger = logging_setup(3)
+
+    pack = repo.create_pack('Pack1')
+    content_object_pack = Pack(pack.path)
+    signer_path = Path('./signer')
+
+    fake_process.register_subprocess(
+        f'{signer_path} {pack.path} keyfile base64', stderr=["error"]
+    )
+
+    content_object_pack.sign_pack(content_object_pack.path, signer_path, 'key')
+
+    captured = capsys.readouterr()
+    assert 'Failed to sign pack for Pack1 -' in captured.err
+
+
+def test_sign_pack_success(repo, capsys, fake_process):
+    """
+    When:
+        - Signing a pack.
+
+    Given:
+        - Pack object.
+
+    Then:
+        - Verify that success is written to the logger.
+
+    """
+    import demisto_sdk.commands.common.content.objects.pack_objects.pack as pack_class
+    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
+
+    pack_class.logger = logging_setup(3)
+
+    pack = repo.create_pack('Pack1')
+    content_object_pack = Pack(pack.path)
+    signer_path = Path('./signer')
+
+    fake_process.register_subprocess(
+        f'{signer_path} {pack.path} keyfile base64', stdout=["success"]
+    )
+
+    content_object_pack.sign_pack(content_object_pack.path, signer_path, 'key')
+
+    captured = capsys.readouterr()
+    assert f'Signed {content_object_pack.path.name} pack successfully' in captured.err
+
+
+def test_encrypt_pack_exception_thrown(repo, capsys, fake_process):
+    """
+    When:
+        - Encrypting a pack.
+
+    Given:
+        - Pack object.
+        - Exception thrown.
+
+    Then:
+        - Verify that exceptions are written to the logger.
+
+    Notes:
+        fake_process fixture will raise the exception in this case - do not delete it.
+
+    """
+    import demisto_sdk.commands.common.content.objects.pack_objects.pack as pack_class
+    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
+
+    pack_class.logger = logging_setup(3)
+
+    pack = repo.create_pack('Pack1')
+    content_object_pack = Pack(pack.path)
+    encryptor_path = Path('./encryptor')
+
+    content_object_pack.encrypt_pack(content_object_pack.path, encryptor_path, 'key')
+
+    captured = capsys.readouterr()
+    assert 'Error while trying to encrypt pack Pack1.' in captured.err
+
+
+def test_encrypt_pack_error_from_subprocess(repo, capsys, fake_process, mocker):
+    """
+    When:
+        - Encrypting a pack.
+
+    Given:
+        - Pack object.
+        - subprocess is failing due to an error.
+
+    Then:
+        - Verify that exceptions are written to the logger.
+
+    """
+    import demisto_sdk.commands.common.content.objects.pack_objects.pack as pack_class
+    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
+
+    pack_class.logger = logging_setup(3)
+
+    pack = repo.create_pack('Pack1')
+    content_object_pack = Pack(pack.path)
+    encryptor_path = Path('./encryptor')
+
+    fake_process.register_subprocess(
+        f'{encryptor_path} {pack.path} {pack.path} "key"', stderr=["error"]
+    )
+
+    content_object_pack.encrypt_pack(content_object_pack.path, encryptor_path, 'key')
+
+    captured = capsys.readouterr()
+    assert 'Failed to encrypt pack for Pack1 -' in captured.err
+
+
+def test_encrypt_pack_success(repo, capsys, fake_process, mocker):
+    """
+    When:
+        - Encrypting a pack.
+
+    Given:
+        - Pack object.
+
+    Then:
+        - Verify that success is written to the logger.
+
+    """
+    import demisto_sdk.commands.common.content.objects.pack_objects.pack as pack_class
+    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
+    import os
+
+    mocker.patch.object(os, 'remove', return_value=None)
+
+    pack_class.logger = logging_setup(3)
+
+    pack = repo.create_pack('Pack1')
+    content_object_pack = Pack(pack.path)
+    encryptor_path = Path('./encryptor')
+
+    fake_process.register_subprocess(
+        f'{encryptor_path} {pack.path} {pack.path} "key"', stdout=['success']
+    )
+
+    content_object_pack.encrypt_pack(content_object_pack.path, encryptor_path, 'key')
+
+    captured = capsys.readouterr()
+    assert f'Encrypted {content_object_pack.path.name} pack successfully' in captured.err
