+++ demisto_sdk/commands/create_artifacts/tests/content_artifacts_creator_test.py
@@ -1,6 +1,4 @@
-import os
 from contextlib import contextmanager
-from datetime import datetime
 from filecmp import cmp, dircmp
 from pathlib import Path
 from shutil import copyfile, copytree, rmtree
@@ -9,7 +7,6 @@ import pytest
 from demisto_sdk.commands.common.constants import PACKS_DIR, TEST_PLAYBOOKS_DIR
 from demisto_sdk.commands.common.logger import logging_setup
 from demisto_sdk.commands.common.tools import src_root
-from packaging.version import parse
 from TestSuite.test_tools import ChangeCWD
 
 TEST_DATA = src_root() / 'tests' / 'test_files'
@@ -241,235 +238,6 @@ def test_duplicate_file_failure(mock_git):
     assert exit_code == 1
 
 
-def test_load_user_metadata_basic(repo):
-    """
-    When:
-        - Dumping a specific pack, processing the pack's metadata.
-
-    Given:
-        - Pack object.
-
-    Then:
-        - Verify that pack's metadata information was loaded successfully.
-
-    """
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (
-        ArtifactsManager, load_user_metadata)
-
-    pack_1 = repo.setup_one_pack('Pack1')
-    pack_1.pack_metadata.write_json(
-        {
-            'name': 'Pack Number 1',
-            'description': 'A description for the pack',
-            'created': '2020-06-08T15:37:54Z',
-            'price': 0,
-            'support': 'xsoar',
-            'url': 'some url',
-            'email': 'some email',
-            'currentVersion': '1.1.1',
-            'author': 'Cortex XSOAR',
-            'tags': ['tag1'],
-            'dependencies': [{'dependency': {'dependency': '1'}}]
-        }
-    )
-
-    with ChangeCWD(repo.path):
-        with temp_dir() as temp:
-            artifact_manager = ArtifactsManager(artifacts_path=temp,
-                                                content_version='6.0.0',
-                                                zip=False,
-                                                suffix='',
-                                                cpus=1,
-                                                packs=True)
-
-    result = load_user_metadata(artifact_manager.content.packs['Pack1'])
-    assert result.id == 'Pack1'
-    assert result.name == 'Pack Number 1'
-    assert result.description == 'A description for the pack'
-    assert result.created == datetime(2020, 6, 8, 15, 37, 54)
-    assert result.price == 0
-    assert result.support == 'xsoar'
-    assert result.url == 'some url'
-    assert result.email == 'some email'
-    assert result.certification == 'certified'
-    assert result.current_version == parse('1.1.1')
-    assert result.author == 'Cortex XSOAR'
-    assert result.tags == ['tag1']
-    assert result.dependencies == [{'dependency': {'dependency': '1'}}]
-
-
-def test_load_user_metadata_advanced(repo):
-    """
-    When:
-        - Dumping a specific pack, processing the pack's metadata.
-
-    Given:
-        - Pack object.
-
-    Then:
-        - Verify that pack's metadata information was loaded successfully.
-
-    """
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (
-        ArtifactsManager, load_user_metadata)
-
-    pack_1 = repo.setup_one_pack('Pack1')
-    pack_1.pack_metadata.write_json(
-        {
-            'name': 'Pack Number 1',
-            'price': 10,
-            'tags': ['tag1'],
-            'useCases': ['usecase1'],
-            'vendorId': 'vendorId',
-            'vendorName': 'vendorName'
-        }
-    )
-
-    with ChangeCWD(repo.path):
-        with temp_dir() as temp:
-            artifact_manager = ArtifactsManager(artifacts_path=temp,
-                                                content_version='6.0.0',
-                                                zip=False,
-                                                suffix='',
-                                                cpus=1,
-                                                packs=True)
-
-    result = load_user_metadata(artifact_manager.content.packs['Pack1'])
-    assert result.id == 'Pack1'
-    assert result.name == 'Pack Number 1'
-    assert result.price == 10
-    assert result.vendor_id == 'vendorId'
-    assert result.vendor_name == 'vendorName'
-    assert result.tags == ['tag1', 'Use Case']
-
-
-def test_load_user_metadata_no_metadata_file(repo, capsys):
-    """
-    When:
-        - Dumping a pack with no pack_metadata file.
-
-    Given:
-        - Pack object.
-
-    Then:
-        - Verify that exceptions are written to the logger.
-
-    """
-    import demisto_sdk.commands.create_artifacts.content_artifacts_creator as cca
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (
-        ArtifactsManager, load_user_metadata)
-
-    cca.logger = logging_setup(3)
-
-    pack_1 = repo.setup_one_pack('Pack1')
-    pack_1.pack_metadata.write_json(
-        {
-            'name': 'Pack Number 1',
-            'price': 'price',
-            'tags': ['tag1'],
-            'useCases': ['usecase1'],
-            'vendorId': 'vendorId',
-            'vendorName': 'vendorName'
-        }
-    )
-
-    with ChangeCWD(repo.path):
-        os.remove(pack_1.pack_metadata.path)
-        with temp_dir() as temp:
-            artifact_manager = ArtifactsManager(artifacts_path=temp,
-                                                content_version='6.0.0',
-                                                zip=False,
-                                                suffix='',
-                                                cpus=1,
-                                                packs=True)
-
-    load_user_metadata(artifact_manager.content.packs['Pack1'])
-
-    captured = capsys.readouterr()
-    assert 'Pack1 pack is missing pack_metadata.json file.' in captured.err
-
-
-def test_load_user_metadata_invalid_price(repo, capsys):
-    """
-    When:
-        - Dumping a pack with invalid price in pack_metadata file.
-
-    Given:
-        - Pack object.
-
-    Then:
-        - Verify that exceptions are written to the logger.
-
-    """
-    import demisto_sdk.commands.create_artifacts.content_artifacts_creator as cca
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (
-        ArtifactsManager, load_user_metadata)
-
-    cca.logger = logging_setup(3)
-
-    pack_1 = repo.setup_one_pack('Pack1')
-    pack_1.pack_metadata.write_json(
-        {
-            'name': 'Pack Number 1',
-            'price': 'price',
-            'tags': ['tag1'],
-            'useCases': ['usecase1'],
-            'vendorId': 'vendorId',
-            'vendorName': 'vendorName'
-        }
-    )
-
-    with ChangeCWD(repo.path):
-        with temp_dir() as temp:
-            artifact_manager = ArtifactsManager(artifacts_path=temp,
-                                                content_version='6.0.0',
-                                                zip=False,
-                                                suffix='',
-                                                cpus=1,
-                                                packs=True)
-
-    load_user_metadata(artifact_manager.content.packs['Pack1'])
-
-    captured = capsys.readouterr()
-    assert 'Pack Number 1 pack price is not valid. The price was set to 0.' in captured.err
-
-
-def test_load_user_metadata_bad_pack_metadata_file(repo, capsys):
-    """
-    When:
-        - Dumping a pack with invalid pack_metadata file.
-
-    Given:
-        - Pack object.
-
-    Then:
-        - Verify that exceptions are written to the logger.
-
-    """
-    import demisto_sdk.commands.create_artifacts.content_artifacts_creator as cca
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (
-        ArtifactsManager, load_user_metadata)
-
-    cca.logger = logging_setup(3)
-
-    pack_1 = repo.setup_one_pack('Pack1')
-    pack_1.pack_metadata.write_as_text('Invalid of course {')
-
-    with ChangeCWD(repo.path):
-        with temp_dir() as temp:
-            artifact_manager = ArtifactsManager(artifacts_path=temp,
-                                                content_version='6.0.0',
-                                                zip=False,
-                                                suffix='',
-                                                cpus=1,
-                                                packs=True)
-
-    load_user_metadata(artifact_manager.content.packs['Pack1'])
-
-    captured = capsys.readouterr()
-    assert 'Failed loading Pack1 user metadata.' in captured.err
-
-
 @pytest.mark.parametrize('key, tool', [('some_key', False), ('', True)])
 def test_sign_packs_failure(repo, capsys, key, tool):
     """
@@ -513,100 +281,6 @@ def test_sign_packs_failure(repo, capsys, key, tool):
            'sign_directory arguments.' in captured.err
 
 
-def test_sign_pack_exception_thrown(repo, capsys, mocker):
-    """
-    When:
-        - Signing a pack.
-
-    Given:
-        - Pack object.
-        - No signing executable.
-
-    Then:
-        - Verify that exceptions are written to the logger.
-
-    """
-    import demisto_sdk.commands.create_artifacts.content_artifacts_creator as cca
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import sign_pack
-    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
-    import subprocess
-
-    mocker.patch.object(subprocess, 'Popen', autospec=True)
-
-    cca.logger = logging_setup(3)
-
-    pack = repo.create_pack('Pack1')
-    content_object_pack = Pack(pack.path)
-
-    sign_pack(content_object_pack, content_object_pack.path, 'key')
-
-    captured = capsys.readouterr()
-    assert 'Error while trying to sign pack Pack1' in captured.err
-
-
-def test_sign_pack_error_from_subprocess(repo, capsys, fake_process):
-    """
-    When:
-        - Signing a pack.
-
-    Given:
-        - Pack object.
-        - subprocess is failing due to an error.
-
-    Then:
-        - Verify that exceptions are written to the logger.
-
-    """
-    import demisto_sdk.commands.create_artifacts.content_artifacts_creator as cca
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import sign_pack
-    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
-
-    cca.logger = logging_setup(3)
-
-    pack = repo.create_pack('Pack1')
-    content_object_pack = Pack(pack.path)
-
-    fake_process.register_subprocess(
-        f'./signDirectory {pack.path} keyfile base64', stderr=["error"]
-    )
-
-    sign_pack(content_object_pack, content_object_pack.path, 'key')
-
-    captured = capsys.readouterr()
-    assert 'Failed to sign pack for Pack1 -' in captured.err
-
-
-def test_sign_pack_success(repo, capsys, fake_process):
-    """
-    When:
-        - Signing a pack.
-
-    Given:
-        - Pack object.
-
-    Then:
-        - Verify that success is written to the logger.
-
-    """
-    import demisto_sdk.commands.create_artifacts.content_artifacts_creator as cca
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import sign_pack
-    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
-
-    cca.logger = logging_setup(3)
-
-    pack = repo.create_pack('Pack1')
-    content_object_pack = Pack(pack.path)
-
-    fake_process.register_subprocess(
-        f'./signDirectory {pack.path} keyfile base64', stdout=['success']
-    )
-
-    sign_pack(content_object_pack, content_object_pack.path, 'key')
-
-    captured = capsys.readouterr()
-    assert f'Signed {content_object_pack.path.name} pack successfully' in captured.err
-
-
 @pytest.mark.parametrize('key, tool', [('some_key', False), ('', True)])
 def test_encrypt_packs_failure(repo, capsys, key, tool):
     """
@@ -648,133 +322,3 @@ def test_encrypt_packs_failure(repo, capsys, key, tool):
     captured = capsys.readouterr()
     assert 'Failed to encrypt packs. In order to do so, you need to provide both encryption_key and ' \
            'encryptor arguments.' in captured.err
-
-
-def test_encrypt_pack_exception_thrown(repo, capsys):
-    """
-    When:
-        - Encrypting a pack.
-
-    Given:
-        - Pack object.
-        - Exception thrown.
-
-    Then:
-        - Verify that exceptions are written to the logger.
-
-    """
-    import demisto_sdk.commands.create_artifacts.content_artifacts_creator as cca
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import ArtifactsManager, encrypt_pack
-    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
-
-    cca.logger = logging_setup(3)
-
-    pack = repo.create_pack('Pack1')
-    content_object_pack = Pack(pack.path)
-
-    with ChangeCWD(repo.path):
-        with temp_dir() as temp:
-            artifact_manager = ArtifactsManager(artifacts_path=temp,
-                                                content_version='6.0.0',
-                                                zip=False,
-                                                suffix='',
-                                                cpus=1,
-                                                packs=True,
-                                                encryption_key='key')
-
-    encrypt_pack(artifact_manager, content_object_pack, content_object_pack.path, 'key')
-
-    captured = capsys.readouterr()
-    assert 'Error while trying to encrypt pack Pack1.' in captured.err
-
-
-def test_encrypt_pack_error_from_subprocess(repo, capsys, fake_process, mocker):
-    """
-    When:
-        - Encrypting a pack.
-
-    Given:
-        - Pack object.
-        - subprocess is failing due to an error.
-
-    Then:
-        - Verify that exceptions are written to the logger.
-
-    """
-    import demisto_sdk.commands.create_artifacts.content_artifacts_creator as cca
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import ArtifactsManager, encrypt_pack
-    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
-    import os
-
-    mocker.patch.object(os, 'chdir', return_value=None)
-
-    cca.logger = logging_setup(3)
-
-    pack = repo.create_pack('Pack1')
-    content_object_pack = Pack(pack.path)
-
-    fake_process.register_subprocess(
-        f'encryptor_path {pack.path} {pack.path} "key"', stderr=["error"]
-    )
-
-    with ChangeCWD(repo.path):
-        with temp_dir() as temp:
-            artifact_manager = ArtifactsManager(artifacts_path=temp,
-                                                content_version='6.0.0',
-                                                zip=False,
-                                                suffix='',
-                                                cpus=1,
-                                                packs=True,
-                                                encryption_key='key',
-                                                encryptor=Path('encryptor_path'))
-
-    encrypt_pack(artifact_manager, content_object_pack, content_object_pack.path, 'key')
-
-    captured = capsys.readouterr()
-    assert 'Failed to encrypt pack for Pack1 -' in captured.err
-
-
-def test_encrypt_pack_success(repo, capsys, fake_process, mocker):
-    """
-    When:
-        - Encrypting a pack.
-
-    Given:
-        - Pack object.
-
-    Then:
-        - Verify that success is written to the logger.
-
-    """
-    import demisto_sdk.commands.create_artifacts.content_artifacts_creator as cca
-    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import ArtifactsManager, encrypt_pack
-    from demisto_sdk.commands.common.content.objects.pack_objects.pack import Pack
-    import os
-
-    mocker.patch.object(os, 'chdir', return_value=None)
-    mocker.patch.object(os, 'remove', return_value=None)
-
-    cca.logger = logging_setup(3)
-
-    pack = repo.create_pack('Pack1')
-    content_object_pack = Pack(pack.path)
-
-    fake_process.register_subprocess(
-        f'encryptor_path {pack.path} {pack.path} "key"', stdout=['success']
-    )
-
-    with ChangeCWD(repo.path):
-        with temp_dir() as temp:
-            artifact_manager = ArtifactsManager(artifacts_path=temp,
-                                                content_version='6.0.0',
-                                                zip=False,
-                                                suffix='',
-                                                cpus=1,
-                                                packs=True,
-                                                encryption_key='key',
-                                                encryptor=Path('encryptor_path'))
-
-    encrypt_pack(artifact_manager, content_object_pack, content_object_pack.path, 'key')
-
-    captured = capsys.readouterr()
-    assert f'Encrypted {content_object_pack.path.name} pack successfully' in captured.err
